requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pycparser==2.21
PyJWT==2.3.0
yarl==1.7.2
